refactor: organizar interfaces
