+++ src/data/contracts/cripto/encrypter.ts
