+++ src/data/contracts/cripto/hash-comparer.ts
